ssh-key: factor `KeypairData` into its own module
The `private` module it was previously inside of had grown quite large.
